+++ tested/evaluators/__init__.py
@@ -23,14 +23,12 @@ For example, such a function looks like this:
         pass
 """
 import functools
-from dataclasses import field
 from pathlib import Path
-from typing import Any, Callable, Dict, List, NamedTuple, Optional, Tuple, Union
-
-from pydantic.dataclasses import dataclass
+from typing import Union
 
 from tested.configs import Bundle
-from tested.dodona import Message, Status, StatusMessage
+from tested.dodona import Status
+from tested.evaluators.common import Evaluator, _curry_evaluator
 from tested.testsuite import (
     EmptyChannel,
     ExceptionBuiltin,
@@ -40,57 +38,12 @@ from tested.testsuite import (
     GenericValueEvaluator,
     IgnoredChannel,
     NormalOutputChannel,
-    OutputChannel,
     ProgrammedEvaluator,
     SpecialOutputChannel,
     SpecificEvaluator,
     TextBuiltin,
     ValueBuiltin,
 )
-
-
-@dataclass
-class EvaluationResult:
-    """Provides the result of an evaluation for a specific output channel."""
-
-    result: StatusMessage  # The result of the evaluation.
-    readable_expected: str
-    """
-    A human-friendly version of what the channel should have been.
-    """
-    readable_actual: str
-    """
-    A human-friendly version (on a best-efforts basis) of what the channel is.
-    """
-    messages: List[Message] = field(default_factory=list)
-    is_multiline_string: bool = False
-    """
-    Indicates if the evaluation result is a multiline string
-    """
-
-
-class EvaluatorConfig(NamedTuple):
-    bundle: Bundle
-    options: Dict[str, Any]
-    context_dir: Path
-
-
-RawEvaluator = Callable[[EvaluatorConfig, OutputChannel, str], EvaluationResult]
-
-Evaluator = Callable[[OutputChannel, str], EvaluationResult]
-
-
-def _curry_evaluator(
-    bundle: Bundle,
-    context_dir: Path,
-    function: RawEvaluator,
-    options: Optional[dict] = None,
-) -> Evaluator:
-    if options is None:
-        options = dict()
-    config = EvaluatorConfig(bundle, options, context_dir)
-    # noinspection PyTypeChecker
-    return functools.partial(function, config)
 
 
 def get_evaluator(
@@ -152,15 +105,3 @@ def get_evaluator(
         return currier(specific.evaluate)
     else:
         raise AssertionError(f"Unknown evaluator type: {type(evaluator)}")
-
-
-def try_outputs(
-    actual: str, parsers: List[Callable[[str], Tuple[Optional[str], Optional[Message]]]]
-) -> Tuple[str, Optional[Message]]:
-    if not actual:
-        return actual, None
-    for parser in parsers:
-        possible, msg = parser(actual)
-        if possible is not None:
-            return possible, msg
-    return actual, None
