+++ tested/evaluators/nothing.py
@@ -4,13 +4,9 @@ RawEvaluator for channels without output.
 import functools
 
 from tested.dodona import Status, StatusMessage
-from tested.evaluators import (
-    EvaluationResult,
-    EvaluatorConfig,
-    exception,
-    try_outputs,
-    value,
-)
+from tested.evaluators.common import EvaluationResult, EvaluatorConfig, try_outputs
+from tested.evaluators.exception import try_as_readable_exception
+from tested.evaluators.value import try_as_readable_value
 from tested.internationalization import get_i18n_string
 from tested.testsuite import EmptyChannel
 
@@ -26,8 +22,8 @@ def evaluate(
 
     if actual:
         parsers = [
-            functools.partial(exception.try_as_readable_exception, config),
-            functools.partial(value.try_as_readable_value, config.bundle),
+            functools.partial(try_as_readable_exception, config),
+            functools.partial(try_as_readable_value, config.bundle),
         ]
         actual, msg = try_outputs(actual, parsers)
         if msg:
