+++ tested/evaluators/value.py
@@ -13,7 +13,7 @@ from tested.datatypes import (
     SimpleTypes,
 )
 from tested.dodona import ExtendedMessage, Permission, Status, StatusMessage
-from tested.evaluators import EvaluationResult, EvaluatorConfig
+from tested.evaluators.common import EvaluationResult, EvaluatorConfig
 from tested.internationalization import get_i18n_string
 from tested.languages.config import TypeSupport
 from tested.languages.generator import convert_statement
